Improve testing and debuging
Checks
------

    Test with rocky versions of oslo's packages
    Test with train versions of oslo's packages
    Test with python 2.7
    Test with python 3.5

No matter the oslo's packages versions in use for this scenario all tests failed.
No matter the python's versions in use in this scenario all tests failed.

Footnote
--------

I just sharing these changes with you to observe together possible path to
follow in an interactive fashion.

I'll try to write a test which will upgrade each oslo packages one by one
until the error is isolated and if we get the right upgrade which could
introduce the issue then I'll try to bisect with git to find the culprit.

However I thing it could be interesting to do the same things with non oslo
packages while we can't get a version which works fine no matter the oslo's
packages versions (rocky or train).

Also it could be worth to check with stein's versions too.

Thoughts?